chore: logic of onDelete Cascade and dynamic model

+++ src/actions/delete/category/index.ts
@@ -1,20 +1,63 @@
 'use server'
 
-import { PrismaClient } from '@prisma/client'
+import { OrderStatus, OrderStatusTracking, PrismaClient } from '@prisma/client'
 
 const prisma = new PrismaClient()
 
 export const deleteCategory = async (categoryId: string) => {
   try {
-    await prisma.category.delete({
+    const ordersProduct = await prisma.orderProduct.findMany({
       where: {
-        id: categoryId,
+        product: {
+          categoryId,
+        },
+      },
+      include: {
+        order: true,
       },
     })
 
-    return { message: 'Categoria deletada.' }
+    const hasOrderProductWithExpectedStatus = ordersProduct.find(
+      (orderProduct) =>
+        (orderProduct.order.orderTracking === OrderStatusTracking.WAITING ||
+          orderProduct.order.orderTracking ===
+            OrderStatusTracking.PRODUCT_DELIVERED_TO_CORREIOS) &&
+        orderProduct.order.status === OrderStatus.PAYMENT_CONFIRMED,
+    )
+
+    if (!hasOrderProductWithExpectedStatus) {
+      await prisma.category.delete({
+        where: {
+          id: categoryId,
+        },
+      })
+
+      const countOrderProducts = await prisma.order.findUnique({
+        where: {
+          id: ordersProduct[0].orderId,
+        },
+        include: {
+          orderProducts: true,
+        },
+      })
+
+      if (countOrderProducts?.orderProducts.length === 0) {
+        await prisma.order.delete({
+          where: {
+            id: countOrderProducts.id,
+          },
+        })
+      }
+
+      return { messageSuccess: 'Categoria deletada.' }
+    } else {
+      return {
+        messageError:
+          'Produtos atualmente em processo com clientes estão associados a esta categoria e não podem ser removidos no momento.',
+      }
+    }
   } catch (err) {
     console.log(err)
-    return { message: 'Error ao deletar categoria.' }
+    return { messageError: 'Error ao deletar categoria.' }
   }
 }
